Move `frame-util` functions to other files
Moved `frame-util` functionality to `frame-observer.js` and
`hypothesis-injector.js`:

- `isAccessible`, `findFrames` and `isDocumentReady` functions to
  `frame-observer.js`, and

- `hasHypothesis` and `injectHypothesis` functions to
  `hypothesis-injector.js`

Renamed and move `src/annotator/util/test/frame-util-test.js` to
`src/annotator/test/frame-observer-test.js`. This unit test only
exercise the `findFrames` function.
