+++ src/annotator/hypothesis-injector.js
@@ -1,5 +1,4 @@
 import FrameObserver from './frame-observer';
-import * as frameUtil from './util/frame-util';
 
 /**
  * @typedef {import('../shared/bridge').Bridge} Bridge
@@ -50,7 +49,7 @@ export class HypothesisInjector {
    * @param {HTMLIFrameElement} frame
    */
   _addHypothesis(frame) {
-    if (frameUtil.hasHypothesis(frame)) {
+    if (hasHypothesis(frame)) {
       return;
     }
 
@@ -63,7 +62,7 @@ export class HypothesisInjector {
     };
 
     const { clientUrl } = this._config;
-    frameUtil.injectHypothesis(frame, clientUrl, injectedConfig);
+    injectHypothesis(frame, clientUrl, injectedConfig);
   }
 
   /**
@@ -73,4 +72,38 @@ export class HypothesisInjector {
     this._bridge.call('destroyFrame', this._frameIdentifiers.get(frame));
     this._frameIdentifiers.delete(frame);
   }
+}
+
+/**
+ * Check if the Hypothesis client has already been injected into an iframe
+ *
+ * @param {HTMLIFrameElement} iframe
+ */
+function hasHypothesis(iframe) {
+  const iframeWindow = /** @type {Window} */ (iframe.contentWindow);
+  return '__hypothesis' in iframeWindow;
+}
+
+/**
+ * Inject the client's boot script into the iframe. The iframe must be from the
+ * same origin as the current window.
+ *
+ * @param {HTMLIFrameElement} iframe
+ * @param {string} scriptSrc
+ * @param {Record<string, any>} config
+ */
+function injectHypothesis(iframe, scriptSrc, config) {
+  const configElement = document.createElement('script');
+  configElement.className = 'js-hypothesis-config';
+  configElement.type = 'application/json';
+  configElement.innerText = JSON.stringify(config);
+
+  const embedElement = document.createElement('script');
+  embedElement.className = 'js-hypothesis-embed';
+  embedElement.async = true;
+  embedElement.src = scriptSrc;
+
+  const iframeDocument = /** @type {Document} */ (iframe.contentDocument);
+  iframeDocument.body.appendChild(configElement);
+  iframeDocument.body.appendChild(embedElement);
 }
